requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.2
ansi2html==1.8.0
appdirs==1.4.4
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
ccimport==0.3.7
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
colour-demosaicing==0.2.2
colour-science==0.4.1
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.3
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
docker-pycreds==0.4.0
easydict==1.10
efficientnet-pytorch==0.7.1
einops==0.6.1
exceptiongroup==1.1.3
execnet==2.0.2
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
fire==0.5.0
Flask==2.2.5
fonttools==4.42.1
fsspec==2023.9.0
gitdb==4.0.10
GitPython==3.1.36
huggingface-hub==0.17.1
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
imutils==0.5.4
iniconfig==2.0.0
ipython==8.12.2
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lark==1.1.7
lazy_loader==0.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mmcv-full==1.7.1
model-index==0.1.11
MultiScaleDeformableAttention==1.0
nbformat==5.7.0
nest-asyncio==1.5.7
networkx==3.1
ninja==1.11.1
numpy==1.24.4
open3d==0.17.0
opencv-python==4.8.0.76
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.25
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pccm==0.3.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.16.1
pluggy==1.3.0
portalocker==2.7.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.11.1
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.16.1
pyparsing==3.1.1
pyquaternion==0.9.9
pytest==7.4.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
pytorch-metric-learning==1.7.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.28.2
retrying==1.3.4
rich==13.4.2
rpds-py==0.10.3
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
sentry-sdk==1.31.0
setproctitle==1.3.2
SharedArray==3.2.3
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.7
tomli==2.0.1
tqdm==4.65.2
traitlets==5.10.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.10
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.9
yapf==0.40.1
zipp==3.16.2
